% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kanban.R
\name{getSelectedCard}
\alias{getSelectedCard}
\title{Get the Selected Card Data}
\usage{
getSelectedCard(outputId, session = NULL)
}
\arguments{
\item{outputId}{A character string specifying the ID of the Kanban output.}
}
\value{
A list with the selected card's details as a list (listName, title, id, position, clickCount)
}
\description{
Retrieves the details of a card that was clicked on the Kanban board.
}
