% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kanban.R
\name{kanban}
\alias{kanban}
\title{Create a Kanban Board Widget}
\usage{
kanban(
  data,
  styleOptions = list(headerBg = "#fff", headerBgHover = "#fff", headerColor = "#353535",
    headerFontSize = "1rem", listNameFontSize = "1rem", cardTitleFontSize = "1rem",
    cardTitleFontWeight = 600, cardSubTitleFontSize = "0.8rem", cardSubTitleFontWeight =
    300, addCardBgColor = "#999", deleteList = list(backgroundColor = "#fff", color =
    "#353535", icon = bsicons::bs_icon("x"), size = "1rem"), deleteCard =
    list(backgroundColor = "#fff", color = "#353535", icon = bsicons::bs_icon("trash"),
    size = "1rem"), addButtonText = "Add", 
     cancelButtonText = "Cancel",
    addCardButtonText = "Add Card", cancelCardButtonText = "Cancel"),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{A named list representing the board data.}

\item{styleOptions}{A named list of style options.}

\item{width, height}{Optional widget dimensions.}

\item{elementId}{DOM element ID.}
}
\description{
This function creates an interactive Kanban board as an HTML widget.
}
