% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kanban.R
\name{updateKanban}
\alias{updateKanban}
\title{Update the data for a Kanban input on the client.}
\usage{
updateKanban(session, inputId, data)
}
\arguments{
\item{session}{The Shiny session object.}

\item{inputId}{The ID of the input object.}

\item{data}{The data to set.}
}
\value{
None
}
\description{
Update the data for a Kanban input on the client.
}
